feat(medecins): correction class and widget
